Refactor of regenerators
* Changed of location the component and layout generators
* Deleted route generator because we don't need them anymore, all routes are defined in /shared/lib/routes.js
